pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "facebook-posts-analysis"
version = "0.1.0"
description = "Local-first pipeline for collecting Facebook page posts/comments and analyzing narratives and stance."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "duckdb>=1.1.3,<2.0.0",
  "hdbscan>=0.8.39,<1.0.0",
  "httpx>=0.27.2,<1.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "langdetect>=1.0.9,<2.0.0",
  "markdown>=3.7,<4.0.0",
  "numpy>=1.26.4,<3.0.0",
  "playwright>=1.48.0,<2.0.0",
  "polars>=1.9.0,<2.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "pydantic>=2.9.2,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "typer>=0.12.5,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3,<9.0.0"
]

[project.scripts]
facebook-posts-analysis = "facebook_posts_analysis.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/facebook_posts_analysis"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

